:hammer: Refactor: diary util함수 src/utils로 이동
